requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==2017.7.27.1
chardet==3.0.4
configobj==4.7.2
coreapi==2.3.1
coreschema==0.0.4
decorator==3.4.0
Django==1.11.3
django-grappelli==2.10.1
django-redis==4.8.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
drfdocs==0.0.11
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
idna==2.5
iniparse==0.4
itypes==1.1.0
Jinja2==2.9.6
kitchen==1.1.1
MarkupSafe==1.0
meld3==1.0.2
MySQL-python==1.2.5
openapi-codec==1.3.2
perf==0.1
pycurl==7.19.0
PyJWT==1.5.2
pyliblzma==0.5.3
python-augeas==0.5.0
pytz==2017.2
redis==2.10.5
requests==2.18.2
simplejson==3.11.1
six==1.9.0
supervisor==3.3.3
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
